+++ client/CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Added events relationship to sets ([#88](https://github.com/project-lux/lux-frontend/issues/88)).
+- Added Professional Activity to facets ([#268](https://github.com/project-lux/lux-frontend/issues/268)).
 
 ### Changed